+++ jdtls.ext/com.microsoft.java.maven.plugin/src/main/java/com/microsoft/java/maven/AddDependencyHandler.java
@@ -28,11 +28,8 @@ import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
 import org.eclipse.lsp4j.TextEdit;
 import org.eclipse.lsp4j.WorkspaceEdit;
-import org.xml.sax.Attributes;
-import org.xml.sax.Locator;
 import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
-import org.xml.sax.helpers.DefaultHandler;
 
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParser;
@@ -56,7 +53,7 @@ public class AddDependencyHandler {
         final int character;
         final int length;
 
-        public AddDependencyParams(String fullClassName, String artifactInfo, 
+        public AddDependencyParams(String fullClassName, String artifactInfo,
                 String uri, int line, int character, int length) {
             this.fullClassName = fullClassName;
             this.artifactInfo = artifactInfo;
@@ -67,7 +64,7 @@ public class AddDependencyHandler {
         }
     }
 
-    private static WorkspaceEdit importEdit(AddDependencyParams params, 
+    private static WorkspaceEdit importEdit(AddDependencyParams params,
             ICompilationUnit unit, CompilationUnit astRoot) throws CoreException {
         // import the new class
         final CompilationUnitChange cuChange = new CompilationUnitChange("", unit);
@@ -77,13 +74,13 @@ public class AddDependencyHandler {
         return ChangeUtil.convertToWorkspaceEdit(cuChange);
     }
 
-    private static WorkspaceEdit replaceEdit(AddDependencyParams params, 
+    private static WorkspaceEdit replaceEdit(AddDependencyParams params,
             ICompilationUnit unit) throws JavaModelException {
         // replace old class name with new one
         final Map<String, List<TextEdit>> textEdits = new HashMap<>();
         textEdits.put(unit.getCorrespondingResource().getLocationURI().toString(),
-                    Arrays.asList(new TextEdit(new Range(new Position(params.line, params.character), 
-                    new Position(params.line, params.character + params.length)), 
+                    Arrays.asList(new TextEdit(new Range(new Position(params.line, params.character),
+                    new Position(params.line, params.character + params.length)),
                     params.fullClassName.substring(params.fullClassName.lastIndexOf('.') + 1))));
         return new WorkspaceEdit(textEdits);
     }
@@ -106,7 +103,7 @@ public class AddDependencyHandler {
         }
     }
 
-    private static WorkspaceEdit pomEdit(AddDependencyParams params, 
+    private static WorkspaceEdit pomEdit(AddDependencyParams params,
             ICompilationUnit unit) throws JavaModelException {
         final IPath path = unit.getJavaProject().getCorrespondingResource().getLocation().append("pom.xml");
         final String info[] = params.artifactInfo.replaceAll(" ", "").split(":");
@@ -114,7 +111,7 @@ public class AddDependencyHandler {
             return new WorkspaceEdit();
         }
         final PosInfo posInfo = getPosInfo(path.toString(), info[0] + ":" + info[1]);
-        final String linesep = System.lineSeparator(); 
+        final String linesep = System.lineSeparator();
         String newtext = "";
         final String pomUriString = unit.getJavaProject().getCorrespondingResource()
                 .getLocationURI().toString() + "/pom.xml";
@@ -125,24 +122,24 @@ public class AddDependencyHandler {
                 textEdits.put(pomUriString, new ArrayList<TextEdit>());
             } else if (posInfo.alreadyHasDependencies == false) {
                 final int space = 2;
-                newtext = linesep + StringUtils.repeat(" ", space) + "<dependencies>" + linesep + 
-                        StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep + 
-                        StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + 
-                        linesep + StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + 
-                        "</artifactId>" + linesep + StringUtils.repeat(" ", space + 4) + "<version>" + 
-                        info[2] + "</version>" + linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + 
+                newtext = linesep + StringUtils.repeat(" ", space) + "<dependencies>" + linesep +
+                        StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep +
+                        StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" +
+                        linesep + StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] +
+                        "</artifactId>" + linesep + StringUtils.repeat(" ", space + 4) + "<version>" +
+                        info[2] + "</version>" + linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" +
                         linesep + StringUtils.repeat(" ", space) + "</dependencies>" + linesep;
-                textEdits.put(pomUriString, 
+                textEdits.put(pomUriString,
                     Arrays.asList(new TextEdit(new Range(posInfo.pos, posInfo.pos), newtext)));
             } else {
                 final int space = posInfo.pos.getCharacter();
-                newtext = linesep + StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep + 
-                    StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + linesep + 
-                    StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + "</artifactId>" + 
-                    linesep + StringUtils.repeat(" ", space + 4) + "<version>" + info[2] + "</version>" + 
-                    linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + linesep + 
+                newtext = linesep + StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep +
+                    StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + linesep +
+                    StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + "</artifactId>" +
+                    linesep + StringUtils.repeat(" ", space + 4) + "<version>" + info[2] + "</version>" +
+                    linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + linesep +
                     StringUtils.repeat(" ", space);
-                textEdits.put(pomUriString, 
+                textEdits.put(pomUriString,
                     Arrays.asList(new TextEdit(new Range(posInfo.pos, posInfo.pos), newtext)));
             }
             return new WorkspaceEdit(textEdits);
@@ -151,109 +148,11 @@ public class AddDependencyHandler {
         }
     }
 
-    public static List<WorkspaceEdit> addDependency(AddDependencyParams params, 
+    public static List<WorkspaceEdit> addDependency(AddDependencyParams params,
             IProgressMonitor monitor) throws CoreException, JavaModelException {
         final ICompilationUnit unit = JDTUtils.resolveCompilationUnit(params.uri);
         final CompilationUnit astRoot = ASTResolving.createQuickFixAST(unit, null);
-        return Arrays.asList(replaceEdit(params, unit), importEdit(params, unit, astRoot), 
+        return Arrays.asList(replaceEdit(params, unit), importEdit(params, unit, astRoot),
                             pomEdit(params, unit));
     }
-}
-
-class GetPosHandler extends DefaultHandler {
-    private Locator locator;
-    private PosInfo posInfo = new PosInfo();
-    private static final int LINE_OFFSET = 1;
-    private static final int COLUMN_OFFSET = 1;
-    private static final int DEPENDENCIES_LENGTH = 15; // the length of "</dependencies>"
-    private static final int PROJECT_LENGTH = 10; // the length of "</projects>"
-    private String targetDependency;
-
-    List<String> dependenciesList = new ArrayList<>(); 
-    Boolean inDependency = false;
-    Boolean inGroupId = false;
-    Boolean inArtifactId = false;
-    String groupId;
-    String artifactId;
-
-    public PosInfo getPosInfo() {
-        return posInfo;
-    }
-
-    @Override
-    public void setDocumentLocator(Locator locator) {
-        this.locator = locator;
-    }
-
-    @Override
-    public void startElement(String uri, String localName, String qName, Attributes attributes) throws SAXException {
-        switch (qName) {
-            case "dependency":
-                inDependency = true;
-                break;
-            case "groupId":
-                inGroupId = true;
-                break;
-            case "artifactId":
-                inArtifactId = true;
-                break;
-        }
-    }
-
-    @Override
-    public void characters(char[] ch, int start, int length) throws SAXException {
-        final String content = new String(ch, start, length);
-        if (inDependency) {
-            if (inGroupId) {
-                groupId = content;
-            } else if (inArtifactId) {
-                artifactId = content;
-            }
-        }
-    }
-
-    @Override
-    public void endElement(String uri, String localName, String qName) throws SAXException {
-        switch(qName) {
-            case "dependencies":
-                posInfo.alreadyHasDependencies = true;
-                posInfo.pos = new Position(locator.getLineNumber() - LINE_OFFSET, 
-                    locator.getColumnNumber() - COLUMN_OFFSET - DEPENDENCIES_LENGTH);
-                break;
-            case "project":
-                if (posInfo.pos == null){
-                    posInfo.pos = new Position(locator.getLineNumber() - LINE_OFFSET, 
-                        locator.getColumnNumber() - COLUMN_OFFSET - PROJECT_LENGTH);
-                }
-                break;
-            case "dependency":
-                inDependency = false;
-                dependenciesList.add(groupId + ":" + artifactId);
-                groupId = artifactId = "";
-                break;
-            case "groupId":
-                inGroupId = false;
-                break;
-            case "artifactId":
-                inArtifactId = false;
-                break;
-        }
-    }
-
-    @Override
-    public void endDocument() throws SAXException {
-        if (dependenciesList.contains(targetDependency)) {
-            posInfo.needAddDependency = false;
-        }
-    }
-
-    public void setTargetDependency(String targetDependency) {
-        this.targetDependency = targetDependency;
-    }
-}
-
-class PosInfo {
-    Position pos = null;
-    Boolean alreadyHasDependencies = false;
-    Boolean needAddDependency = true;
 }
